requirements.txt
Pillow
smbus
spidev
lgpio
Rpi.GPIO
gpiozero
